+++ lib/__tests__/backends/terraformGoldenFiles/envs/snowflake.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     snowflake = {
       source = "Snowflake-Labs/snowflake"
-      version = "0.55.1"
+      version = "0.64.0"
     }
   }
 }