+++ lib/backends/terraform/terraformGrant.ts
@@ -3,7 +3,7 @@ import {immerable} from 'immer'
 import {NamingConvention} from '../../namingConvention'
 import {
   Grant,
-  GrantKind
+  GrantKind, isSchemaObjectGrant
 } from '../../grants/grant'
 import {Role} from '../../roles/role'
 import {TerraformRole} from './terraformRole'
@@ -21,8 +21,14 @@ export abstract class TerraformGrant implements TerraformResource {
   abstract toRoles: Role[]
   abstract toTerraformRoles: TerraformRole[]
 
+  type = 'TerraformGrant'
+
   protected roleAndRoleResourceStrings(): string[] {
     return this.toRoles.map(r => `"${r.name}"`).concat(this.toTerraformRoles.map(tr => `snowflake_role.${tr.resourceName()}.name`))
   }
 
+}
+
+export function isTerraformGrant(obj: any): obj is TerraformGrant {
+  return 'kind' in obj && obj.type === 'TerraformGrant'
 }