+++ lib/environment.ts
@@ -1,7 +1,7 @@
 import {Deployable} from './deployable'
 
 export class Environment extends Deployable {
-  toObject(): Record<string, unknown> {
+  toRecord(): Record<string, unknown> {
     return {
       name: this.name
     }