+++ lib/grants/schemaObjectGrant.ts
@@ -3,7 +3,7 @@ import {Role} from '../roles/role'
 import {Privilege} from '../privilege'
 import {Grant, GrantType} from './grant'
 
-export const SchemaObjectGrantKinds = ['file_format', 'function', 'procedure', 'sequence', 'stage', 'stream', 'table', 'task', 'view']
+export const SchemaObjectGrantKinds = ['file_format', 'function', 'materialized_view', 'procedure', 'sequence', 'stage', 'stream', 'table', 'task', 'view']
 export type SchemaObjectGrantKind = typeof SchemaObjectGrantKinds[number]
 
 export abstract class SchemaObjectGrant implements Grant {