+++ lib/objects/schema.ts
@@ -5,6 +5,7 @@ import {SchemaAccessRole} from '../roles/schemaAccessRole'
 import {DataAccessLevel} from '../access/dataAccessLevel'
 import {NamingConvention} from '../namingConvention'
 import {FunctionalRole} from '../roles/functionalRole'
+import {VirtualWarehouseAccessLevel} from '../access/virtualWarehouseAccessLevel'
 
 export interface SchemaOptions {
   managedAccess: boolean
@@ -21,7 +22,6 @@ export const defaultSchemaOptions: SchemaOptions = {
 export class Schema implements SchemaOptions {
   name: string
   database: Database
-  tables: Table[]
   access: DataAccess
   managedAccess: boolean
   transient: boolean
@@ -30,7 +30,6 @@ export class Schema implements SchemaOptions {
   constructor(name: string, database: Database, options: SchemaOptions = defaultSchemaOptions) {
     this.name = name
     this.database = database
-    this.tables = []
     this.access = new Map<FunctionalRole, DataAccessLevel>()
     this.managedAccess = options.managedAccess
     this.transient = options.transient
@@ -41,5 +40,20 @@ export class Schema implements SchemaOptions {
     return [DataAccessLevel.Read, DataAccessLevel.ReadWrite, DataAccessLevel.Full].map(
       (level) => new SchemaAccessRole(this, level, namingConvention)
     )
+  }
+
+  toRecord() {
+    const accessRecords = Array(...this.access.entries()).map(([role, level]) => {
+      return {role: role.name, level: DataAccessLevel[level]}
+    })
+    return {
+      name: this.name,
+      options: {
+        transient: this.transient,
+        dataRetentionTimeInDays: this.dataRetentionTimeInDays,
+        managedAccess: this.managedAccess
+      },
+      access: accessRecords
+    }
   }
 }